moved the intefaces into model and removed the folder
