+++ app/src/main/java/com/goayo/debtify/MockDatabase/MockDatabase.java
@@ -11,7 +11,7 @@ import com.goayo.debtify.model.LoginException;
 import com.goayo.debtify.model.RegistrationException;
 import com.goayo.debtify.model.UserAlreadyExistsException;
 import com.goayo.debtify.model.UserNotFoundException;
-import com.goayo.debtify.modelaccess.IUserData;
+import com.goayo.debtify.model.IUserData;
 import com.google.gson.Gson;
 
 import java.math.BigDecimal;
