+++ app/src/main/java/com/goayo/debtify/model/Account.java
@@ -1,8 +1,5 @@
 package com.goayo.debtify.model;
 
-import com.goayo.debtify.modelaccess.IGroupData;
-import com.goayo.debtify.modelaccess.IUserData;
-
 import java.math.BigDecimal;
 import java.net.ConnectException;
 import java.util.HashMap;
