+++ app/src/main/java/com/goayo/debtify/model/Ledger.java
@@ -2,7 +2,6 @@ package com.goayo.debtify.model;
 
 
 import com.goayo.debtify.Tuple;
-import com.goayo.debtify.modelaccess.IDebtData;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
