+++ app/src/main/java/com/goayo/debtify/view/AddContactFragment.java
@@ -22,7 +22,7 @@ import com.goayo.debtify.R;
 import com.goayo.debtify.databinding.AddContactFragmentBinding;
 import com.goayo.debtify.model.UserAlreadyExistsException;
 import com.goayo.debtify.model.UserNotFoundException;
-import com.goayo.debtify.modelaccess.IUserData;
+import com.goayo.debtify.model.IUserData;
 import com.goayo.debtify.view.adapter.UserCardViewAdapter;
 import com.goayo.debtify.viewmodel.AddContactViewModel;
