+++ app/src/main/java/com/goayo/debtify/view/CreateGroupFragment.java
@@ -17,7 +17,7 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.goayo.debtify.R;
 import com.goayo.debtify.databinding.CreateGroupFragmentBinding;
-import com.goayo.debtify.modelaccess.IUserData;
+import com.goayo.debtify.model.IUserData;
 import com.goayo.debtify.view.adapter.UserCardViewAdapter;
 import com.goayo.debtify.viewmodel.GroupCreationViewModel;
 import com.goayo.debtify.viewmodel.PickUserViewModel;
