+++ app/src/main/java/com/goayo/debtify/view/DisplayContactsFragment.java
@@ -18,7 +18,7 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.goayo.debtify.R;
 import com.goayo.debtify.databinding.DisplayContactsFragmentBinding;
-import com.goayo.debtify.modelaccess.IUserData;
+import com.goayo.debtify.model.IUserData;
 import com.goayo.debtify.view.adapter.UserCardViewAdapter;
 import com.goayo.debtify.viewmodel.ContactsViewModel;
 import com.goayo.debtify.viewmodel.PickUserViewModel;
