+++ app/src/main/java/com/goayo/debtify/view/GroupFragment.java
@@ -20,8 +20,8 @@ import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 import com.goayo.debtify.R;
 import com.goayo.debtify.databinding.GroupFragmentBinding;
 import com.goayo.debtify.model.UserNotFoundException;
-import com.goayo.debtify.modelaccess.IDebtData;
-import com.goayo.debtify.modelaccess.IUserData;
+import com.goayo.debtify.model.IDebtData;
+import com.goayo.debtify.model.IUserData;
 import com.goayo.debtify.view.adapter.TransactionCardAdapter;
 import com.goayo.debtify.viewmodel.DetailedGroupViewModel;
 import com.goayo.debtify.viewmodel.PickUserViewModel;
