+++ app/src/main/java/com/goayo/debtify/view/PickUsersFragment.java
@@ -22,7 +22,7 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.goayo.debtify.R;
 import com.goayo.debtify.databinding.PickUsersFragmentBinding;
-import com.goayo.debtify.modelaccess.IUserData;
+import com.goayo.debtify.model.IUserData;
 import com.goayo.debtify.view.adapter.PickUserAdapter;
 import com.goayo.debtify.viewmodel.PickUserViewModel;
