+++ app/src/main/java/com/goayo/debtify/view/adapter/PickUserAdapter.java
@@ -11,7 +11,7 @@ import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.goayo.debtify.R;
-import com.goayo.debtify.modelaccess.IUserData;
+import com.goayo.debtify.model.IUserData;
 
 import java.util.ArrayList;
 import java.util.List;
