+++ app/src/main/java/com/goayo/debtify/viewmodel/GroupCreationViewModel.java
@@ -3,7 +3,7 @@ package com.goayo.debtify.viewmodel;
 import androidx.lifecycle.ViewModel;
 
 import com.goayo.debtify.model.ModelEngine;
-import com.goayo.debtify.modelaccess.IUserData;
+import com.goayo.debtify.model.IUserData;
 
 import java.util.ArrayList;
 import java.util.HashSet;
