+++ app/src/test/java/com/goayo/debtify/model/AccountTest.java
@@ -1,16 +1,9 @@
 package com.goayo.debtify.model;
 
 import com.goayo.debtify.Database.RealDatabase;
-import com.goayo.debtify.modelaccess.IGroupData;
 
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
-
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
