+++ app/src/test/java/com/goayo/debtify/model/DebtTrackerTest.java
@@ -1,7 +1,5 @@
 package com.goayo.debtify.model;
 
-import com.goayo.debtify.modelaccess.IPaymentData;
-
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
