+++ app/src/test/java/com/goayo/debtify/model/GroupTest.java
@@ -1,14 +1,9 @@
 package com.goayo.debtify.model;
 
-import com.goayo.debtify.modelaccess.IGroupData;
-import com.goayo.debtify.modelaccess.IUserData;
-
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 import static org.junit.Assert.assertEquals;
